# src/docs/swagger.yaml
basePath: /v1cd
...
host: petstore.swagger.io
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API12222
  version: "1.0"
paths:
...
  /hello/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      security:
      - Bearer: []
      summary: Print
      tags:
      - Name
...
swagger: "2.0"